checkpoint

+++ sdk/python/feast/infra/online_stores/contrib/mysql_online_store/util.py
@@ -12,11 +12,12 @@ from pydantic import StrictStr
 from pymysql.connections import Connection
 from pymysql.cursors import Cursor
 
-from feast import FeatureView
+from feast import FeatureView, StreamFeatureView
 from feast.repo_config import FeastConfigBaseModel
 
 from feast.infra.online_stores.contrib.mysql_online_store.defs import (
-    MYSQL_DEADLOCK_ERR
+    MYSQL_DEADLOCK_ERR,
+    FEATURE_VIEW_VERSION_TABLE_NAME
 )
 from feast.infra.online_stores.exceptions import OnlineStoreError
 from feast.protos.feast.types.Value_pb2 import Value as ValueProto
@@ -37,18 +38,16 @@ class MySQLOnlineStoreConfig(FeastConfigBaseModel):
     port: Optional[int] = None
     session_manager_module: Optional[StrictStr] = None
 
-    feature_version_table: str = ''
-    feature_data_table: str = ''
 
-
-def table_id(project: str, table: FeatureView) -> str:
-    return f"{project}_{table.name}"
+def table_id(project: str, fv_name: str) -> str:
+    return f"{project}_{fv_name}"
 
 
 def drop_table_and_index(cur: Cursor, conn: Connection, project: str, table: FeatureView) -> None:
-    table_name = table_id(project, table)
+    table_name = table_id(project, table.name)
     try:
-        cur.execute(f"DROP INDEX {table_name}_ek ON {table_name};")
+        if isinstance(table, StreamFeatureView):
+            cur.execute(f"DROP INDEX {table_name}_ek ON {table_name};")
         cur.execute(f"DROP TABLE IF EXISTS {table_name}")
         conn.commit()
     except Exception as e:
@@ -67,30 +66,32 @@ def to_naive_utc(ts: datetime) -> datetime:
         return ts.astimezone(pytz.utc).replace(tzinfo=None)
 
 
-def default_version_id(feature_view_name: str, offset: int = 0) -> int:
-    offset_bstr = (feature_view_name + '_' + str(offset)).encode('utf-8')
-    return int.from_bytes(hashlib.sha256(offset_bstr).digest()[:8], 'little')
+def default_version_id() -> int:
+    # deterministic version_id used when initializing tables for the first time
+    return 0
 
 
 def build_legacy_features_read_query(
     project: str,
-    table: FeatureView,
+    fv_name: str,
     union_offset: Optional[int] = None
 ) -> str:
     base = "SELECT feature_name, value, event_ts" + " " if union_offset is None else f", {union_offset} as __i__ "
-    base += f"FROM {table_id(project, table)} WHERE entity_key = %s"
+    base += f"FROM {table_id(project, fv_name.name)} WHERE entity_key = %s"
     return base
 
 
 def build_feature_view_features_read_query(
-    config: MySQLOnlineStoreConfig,
-    feature_view_name: str,
+    project: str,
+    fv_name: str,
     union_offset: Optional[int] = None
 ) -> str:
-    version_table, data_table = config.feature_version_table, config.feature_version_table
+    version_table = table_id(project=project, fv_name=FEATURE_VIEW_VERSION_TABLE_NAME)
+    data_table = table_id(project=project, fv_name=fv_name)
+
     base = "SELECT feature_name, value, event_ts" + "" if union_offset is None else f", {union_offset} as __i__"
-    base += f" FROM {data_table} WHERE feature_view_name = {feature_view_name} AND entity_id = %s IN "
-    base += f"(SELECT version_id FROM {data_table} WHERE feature_view_name = {feature_view_name});"
+    base += f" FROM {data_table} WHERE entity_key = %s AND version_id IN "
+    base += f"(SELECT version_id FROM {version_table} WHERE feature_view_name = {fv_name});"
     return base
 
 
@@ -174,19 +175,19 @@ def execute_query_with_retry(
 def get_feature_view_version_id(
     cur: Cursor,
     conn: Connection,
-    config: MySQLOnlineStoreConfig,
-    feature_view_name: str,
+    project: str,
+    feature_view: FeatureView,
 ) -> Optional[int]:
-    table, version_id = config.feature_version_table, None
+    table, version_id = table_id(project=project, table=FEATURE_VIEW_VERSION_TABLE_NAME), None
     cur.execute(
         "SELECT version_id FROM %s WHERE feature_view_name = %s;",
-        (table, feature_view_name)
+        (table, feature_view.name)
     )
     record = cur.fetchone()
     if record is not None:
         version_id = record[0]
     else:
-        logging.info(f'FeatureView ({feature_view_name} not found in {table}).')
+        logging.info(f'FeatureView ({feature_view.name} not found in {table}).')
     conn.commit()
     return version_id
 
@@ -194,14 +195,15 @@ def get_feature_view_version_id(
 def insert_feature_view_version(
     cur: Cursor,
     conn: Connection,
-    config: MySQLOnlineStoreConfig,
-    feature_view_name: str,
+    fv_name: str,
     version_id: int,
+    project: str,
     lock_row: bool = False,
     commit: bool = True
 ) -> None:
-    write_query, table = "INSERT INTO %s VALUES (%s, %s, %s,)", config.feature_version_table
-    cur.execute(write_query, (table, feature_view_name, version_id, lock_row,))
+    write_query = "INSERT INTO %s VALUES (%s, %s, %s,)"
+    table = table_id(project=project, table=FEATURE_VIEW_VERSION_TABLE_NAME)
+    cur.execute(write_query, (table, fv_name, version_id, lock_row,))
     if commit:
         conn.commit()
 
@@ -296,14 +298,14 @@ def _update_feature_view_version_id(
 def create_feature_view_partition(
     cur: Cursor,
     conn: Connection,
-    config: MySQLOnlineStoreConfig,
+    fv_table: str,
     version_id: int
 ) -> str:
-    table, partition_name = config.feature_data_table, get_partition_id_from_version(version_id=version_id)
+    partition_name = get_partition_id_from_version(version_id=version_id)
 
     cur.execute(
         "SELECT COUNT(*) FROM information.schema.partitions WHERE TABLE_NAME = %s AND partition_name IS NOT NULL",
-        (table,)
+        (fv_table,)
     )
     result = cur.fetchone()
     table_not_partitioned = result is None or result == 0
@@ -311,7 +313,7 @@ def create_feature_view_partition(
     partition_query = "ALTER TABLE %s"
     partition_query += "PARTITION BY LIST (version_id)(" if table_not_partitioned else "ADD PARTITION ("
     partition_query += "PARTITION `%s` VALUES IN (%s));"
-    cur.execute(partition_query, (table, partition_name, version_id,))
+    cur.execute(partition_query, (fv_table, partition_name, version_id,))
 
     conn.commit()
     return partition_name
@@ -320,15 +322,15 @@ def create_feature_view_partition(
 def drop_feature_view_partitions(
     cur: Cursor,
     conn: Connection,
-    config: MySQLOnlineStoreConfig,
+    fv_table: str,
     version_ids: List[int],
 ) -> None:
     partition_names = [get_partition_id_from_version(version_id=vid) for vid in version_ids]
     partition_list = ', '.join([f"`{partition_name}`" for partition_name in partition_names])
 
-    alter_query, query_values = "ALTER TABLE %s DROP PARTITION %s", (config.feature_data_table, partition_list,)
+    alter_query, query_values = "ALTER TABLE %s DROP PARTITION %s", (fv_table, partition_list,)
     execute_query_with_retry(
-        cur=cur, conn=conn, query=alter_query, values=query_values, propagate_exceptions=True
+        cur=cur, conn=conn, query=alter_query, values=query_values,
     )
 
 
@@ -353,22 +355,63 @@ def unlock_feature_view_version(
         raise e
 
 
-def drop_feature_view(
+def drop_feature_view_version(
     cur: Cursor,
     conn: Connection,
-    config: MySQLOnlineStoreConfig,
+    project: str,
     feature_view_name: str,
     version_id: int
 ) -> None:
     # RB: version_id in this case roughly acts as a "proof of lock" value.
-    query, values = "DELETE FROM %s WHERE feature_view_name = %s", (config.feature_version_table, feature_view_name,)
+    version_table = table_id(project=project, table=FEATURE_VIEW_VERSION_TABLE_NAME)
+    query, values = "DELETE FROM %s WHERE feature_view_name = %s", (version_table, feature_view_name,)
     try:
         execute_query_with_retry(
-            cur=cur, conn=conn, query=query, values=values, propagate_exceptions=True
+            cur=cur, conn=conn, query=query, values=values,
         )
     except Exception as e:
         unlock_feature_view_version(
-            cur=cur, conn=conn, config=config, feature_view_name=feature_view_name, version_id=version_id
+            cur=cur, conn=conn, project=project, feature_view_name=feature_view_name, version_id=version_id
         )
         raise e
 
+
+def build_feature_view_version_table(
+    cur: Cursor,
+    conn: Connection,
+    project: str
+) -> str:
+    table_name = table_id(project=project, table=FEATURE_VIEW_VERSION_TABLE_NAME)
+    try:
+        cur.execute(
+            "CREATE TABLE IF NOT EXISTS %s (feature_view_name VARCHAR(255), version_id BIGINT, version_locked BOOL)",
+            table_name
+        )
+    except Exception as e:
+        conn.rollback()
+        raise e
+    return table_name
+
+
+def build_feature_view_data_table(
+    cur: Cursor,
+    conn: Connection,
+    feature_view: FeatureView,
+    project: str
+) -> str:
+    table_name = table_id(project=project, table=feature_view)
+    try:
+        cur.execute(
+            "CREATE TABLE IF NOT EXISTS %s ("
+            "version_id BIGINT, "
+            "entity_key VARCHAR(255), "
+            "feature_name VARCHAR(255), "
+            "value BLOB, "
+            "event_ts timestamp NULL DEFAULT NULL, "
+            "created_ts timestamp NULL DEFAULT NULL, "
+            "PRIMARY KEY(version_id, entity_key, feature_name))"
+        )
+    except Exception as e:
+        conn.rollback()
+        raise e
+    return table_name
